requirements.txt
# Core Web Frameworks
Flask==2.2.3
fastapi==0.129.0
uvicorn==0.41.0

# Authentication & Security
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.2.1
PyJWT==2.11.0
passlib==1.7.4

# FastAPI Dependencies
pydantic==2.12.5
pydantic_core==2.41.5
python-multipart==0.0.22
starlette==0.52.1

# Utilities
Werkzeug==2.2.2
colorama==0.4.6

# Production Server
gunicorn==21.2.0

# Testing
httpx==0.25.0
pytest==7.4.3
